erofs-utils: lib: keep erofs_init_devices in sync with kernel
Keep erofs_init_devices() in sync with kernel erofs_scan_devices()[1],
which scans the devtable (if any) automatically if sbi->extra_devices is
not explicitly specified.

Also fix the missing le32_to_cpu() when parsing the device slot.  Read
and cache the number of blocks of each device for later use.

[1] https://git.kernel.org/torvalds/c/ba73eadd23d1

Fixes: 0ce853a01123 ("erofs-utils: fuse: add multiple device support")
Signed-off-by: Jingbo Xu <jefflexu@linux.alibaba.com>
Reviewed-by: Gao Xiang <hsiangkao@linux.alibaba.com>
Link: https://lore.kernel.org/r/20230823071517.12303-3-jefflexu@linux.alibaba.com
Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>