erofs-utils: mkfs: support flatdev for multi-blob images
Since kernel commit 8b465fecc35a ("erofs: support flattened block device
for multi-blob images"), the flatdev feature has been introduced to
support mounting multi-blobs container image as a single block device.

To enable this feature, the mapped_blkaddr of each device slot needs to
be set properly to the offset of the device in the flat address space.

The uuid of the source image is used as the corresponding device tag
in rebuild mode.

Signed-off-by: Jingbo Xu <jefflexu@linux.alibaba.com>
Link: https://lore.kernel.org/r/20230915115333.17599-1-jefflexu@linux.alibaba.com
Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>