erofs-utils: lib: oci: allow HTTP connections to registry
Currently, the URL used to send requests to the registry is hardcoded
with "https://". This patch introduces an optional insecure option for
oci in {mkfs,mount}.erofs, enabling registry access via the HTTP
protocol.

Also, this patch refactors the deeply nested logic in the `--oci`
argument parsing in both {mkfs,mount}/main.c.

Signed-off-by: Yifan Zhao <zhaoyifan28@huawei.com>
Link: https://lore.kernel.org/r/20251205103322.82638-1-zhaoyifan28@huawei.com
Signed-off-by: Gao Xiang <hsiangkao@linux.alibaba.com>